좋아요 기능 추가

+++ src/main/java/com/bb/biz/controller/FavoriteController.java
@@ -3,6 +3,7 @@ package com.bb.biz.controller;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
 
 import com.bb.biz.favorite.FavoriteService;
 import com.bb.biz.favorite.FavoriteVO;
@@ -12,15 +13,20 @@ public class FavoriteController {
 	@Autowired
 	private FavoriteService favoriteService;
 	
-	@RequestMapping(value="/insertF.do")
-	public String InsertFavorite(FavoriteVO fVO) {
-		favoriteService.insertFavorite(fVO);
-		return "main.jsp";
-	}
-	@RequestMapping(value="/deleteF.do")
-	public String DelteFavorite(FavoriteVO fVO) {
-		favoriteService.deleteFavorite(fVO);
-		return "main.jsp";
+	@ResponseBody
+	@RequestMapping(value="/favorite.do")
+	public String Favorite(FavoriteVO fVO) {
+		
+		if(favoriteService.selectOneFavorite(fVO)!=null) { // 해당 FID가 있는지 확인
+			
+			favoriteService.deleteFavorite(fVO);
+			return "0";
+		}
+		else {
+			favoriteService.insertFavorite(fVO);
+			return "1";
+		}
+		
 	}
 	
 }
